testingClient/requirements.txt
streamlit
pillow
requests
google-cloud-storage
google-cloud-automl
python-dotenv
